Bug Fix: Fix vulnerability report for older kernels
The analy.py script assumes that it will always run on a kernel
that has the fixes for the vulnerabilities (even if disabled),
that exports the vulnerabilities folder in:
/sys/devices/system/cpu/vulnerabilities (used by lscpu)
and that the Linux distribution has a version of lscpu that
supports the above features.

If this set of conditions is not met, the output of 'make testing'
displays all the output from lscpu, which is not what is desired.
Fix this by reporting 'Unknown' for old systems that do not have
these fixes.